+++ src/got_mouse.py
@@ -10,11 +10,11 @@ from pet import PetAnimState
 from read_parameters import param_dict
 
 class GotMouse(System):
-    GOT_MIN_DURATION = int(param_dict["GOT_MIN_DURATION"]) # In seconds.
-    GOT_MAX_DURATION = int(param_dict["GOT_MAX_DURATION"])
-    GOT_RUN_SPEED = int(param_dict["GOT_RUN_SPEED"])
-    MOUSE_OFFSET_RUN_LEFT = Vector2(int(param_dict["MOUSE_OFFSET_RUN_LEFT_X"]), int(param_dict["MOUSE_OFFSET_RUN_LEFT_Y"]))
-    MOUSE_OFFSET_RUN_RIGHT = Vector2(int(param_dict["MOUSE_OFFSET_RUN_RIGHT_X"]), int(param_dict["MOUSE_OFFSET_RUN_RIGHT_Y"]))
+    GOT_MIN_DURATION = float(param_dict["GOT_MIN_DURATION"]) # In seconds.
+    GOT_MAX_DURATION = float(param_dict["GOT_MAX_DURATION"])
+    GOT_RUN_SPEED = float(param_dict["GOT_RUN_SPEED"])
+    MOUSE_OFFSET_RUN_LEFT = Vector2(float(param_dict["MOUSE_OFFSET_RUN_LEFT_X"]), float(param_dict["MOUSE_OFFSET_RUN_LEFT_Y"]))
+    MOUSE_OFFSET_RUN_RIGHT = Vector2(float(param_dict["MOUSE_OFFSET_RUN_RIGHT_X"]), float(param_dict["MOUSE_OFFSET_RUN_RIGHT_Y"]))
 
     # to be init at the start
     def on_enter(self, pet):
